Update patch set 1
Patch Set 1: Verified-1 Code-Review-1

scAddr is basically always null, and I guarantee this will NRE all over the place. (you should test before submitting to gerrit)

destAddr is also not guaranteed to be in canonical format. ie, it may just be a 7 digit number.

The proper way to do this is to use libphonenumbergoogle or something to parse and process the number. Or send it to google voice to see if the sending fails, and then fall back to regular delivery.

Patch-set: 1