Update patch set 1
Patch Set 1: -Verified

fair enough, if you null check the stuff, and then find that the prefix exists and is NOT +1, i'd be willing to take that patch.

the way you have it written right now, it's pretty much going to always fail to send for every user on the planet. no one +1 prefixes their numbers.

along hte lines of:

if (number != null && number.startsWith("+") && !number.startsWith("+1")) {
   // this is a non-US number. bail.
}

Patch-set: 1