Update patch set 3
Patch Set 3:

So I've finally gotten around to testing this out.

It looks like when it takes the path of not sending through google voice, then it doesn't send at all. For some reason, it won't then send over cellular, like it should.

Investigating why...

I had previously assumed that this code in onStartCommand was meant to return early and give control back to cellular:


        if (null == settings.getString("account", null)) {
            stopSelf();
            return ret;
        }


So I replicated it in the path for this commit:


            if (!canDeliverToAddress(destination)) {
                if (destination == null)
                    destination = "(null)";
                Log.d(LOGTAG, "Sending <" + destination + "> via cellular instead of gvoice.");
                stopSelf();
                return ret;
            }


Unfortunately,  it doesn't seem to work. At all. It turns out that original code was never even being hit, because setting the account to "Disabled" in the settings disables the listener in the first place.

Anybody know the correct way to signify, "no, sorry, pass this on to the normal cellular channels"?

Patch-set: 3
Reviewer: Jason A. Donenfeld <7509@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0