Update patch set 1
Patch Set 1:

Sorry, what is the problem with a user changing that? This is to help Voice+ appropriately understand the phone numbers in the users contact list. If a phone is Libyan, but the users contact list (synced from Google) are entered as Canadian, then it is likely they will have the language set to en_ca, and the country will be assumed to be Canada, and the phone with Google Voice will work with the Canadian contact list.

Am I misunderstanding an assumption of how the country check is supposed to function? My assumption was that it wasn't to prevent any Voice+ / Google Voice use, but so make it simply understand which numbers to assume are routable.

Patch-set: 1