Update patch set 1
Patch Set 1:

Tried to cover the explanation in the commit, but please feel free to critique. I didn't say it, but without this, Voice+ cannot work on a wifi tablet at all, which is perhaps the most important thing. (Google Voice is not on Play Store for these devices, but can be manually installed or can come with a custom ROM.)

The added condition will only ever affect those 1) with a SIM card, but with radio off and wifi on since device boot, and 2) those with wifi-only devices. Both of these people are unserviced atm.

Patch-set: 1
Reviewer: Patrick Connolly <12975@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0