Update patch set 1
Patch Set 1: Code-Review-1

As i said before, this is like a "return true" for this method. If you want to know if the user is in USA/Canada is better to use GPS than the locale for the WiFi only devices (the locale can be changed by a user).

Patch-set: 1