Update patch set 1
Patch Set 1:

Thanks Jorge, but as I said before splitting this contentious change out: The purpose of these country checks isn't to prevent someone from using GVoice from these countries, but to help Voice+ understand their contacts' phone numbers as entered into the address books. The Locale is very likely to match the expected format of their phone numbers, and if not, the ability to change it is a feature.

As as I said, this fix is for a tiny subset of users who either have 1) radio off by default, or 2) devices without radios. Right now, Voice+ cannot function for them, and this is trying to find a way to get it working. If not this, then we need to expand the Voice+ settings to allow country to be set specifically for Voice+, for which your criticism (which I feel is miplaced) would curiously still apply.

Also, people with radio either off, non-existent, or disabled are not person for which "turn on GPS" is a fair solution.

Patch-set: 1